requirements.txt
fastapi
uvicorn
pydantic
httpx
python-multipart
pandas
scikit-learn
crewai
crewai-tools
chromadb
python-dotenv
sentence-transformers